USB: HAL: Hardcoding USB values to make VTS compatiple for non TYPEC
On kernel 5.4, TYPEC is enabled by default in the gki_defconfig.
So for platforms that do not support TYPEC, the VtsHalUsbV1_2
testing is failing at various checks since here the
"/sys/class/typec" is present but is an empty directory.
To make the VtsHalUsbV1_2 compatible with the non TYPEC target
also, hardcode the necessary values so that the various checks in
the VtsHalUsbV1_2TargetTest  passes.

Change-Id: Iecba4c896c5c6ba6c3dbcc739dffe88c2c5c106c