USB: HAL: Add support to toggle mode based on uevent
Add support on usbd to toggle dwc3 mode when hcd reports
host controller error event. The USB hcd puts the host
controller in a halt state when the host fails to respond
hcd commands due to some internal error. The current
implementation does not help recover from this state until
a power cycle is performed, but this changes helps recover
the halted host in the same power cycle.

Change-Id: Icd3db51ea2fa2eb7e15fcf4572de9bd6af362ece
Signed-off-by: Udipto Goswami <quic_ugoswami@quicinc.com>
Signed-off-by: Selvarasu Ganesan <quic_selvaras@quicinc.com>