init.qcom.usb.sh: Add support for control extension descriptors
Add support for uvc extension unit descriptor which will
support control parameters with below configuration,
- directory :control/extensions/xu.0
- baSourceID :4
- bNrInpins :1
- bNumControls :5
- bControlSize :2
- bmControls :0xFF, 0xFF

Here, the parameters are updated as per the Vendor Utility App
requirements, which are not public. The GUID is proprietary to Dell,
so internally needs to be populated at runtime.
We have exposed 15 controls as supported here, which need to change
based on supported controls. SourceID is mapped to UVC Camera. The
rest of the parameters as per UVC spec 1.5.

Change-Id: I62bf7d3e4a2c92b5dee645b491909d9f4259fe72