UsbGadget: Include USB FFS AOA mount points
On Android16, there is a requirement for creation of USB FFS devices for
the AOA protocol.  AOA is planned to be moved from the legacy f_accessory
kernel driver into an userspace entity leveraging USB FFS, but the
transition is occurring late on projects utilizing Android16.  For the time
being, in order to fulfill the requirements, add the FFS nodes, but
continue utilizing the f_accessory driver for AOA.

[GMS-VSR-5.4-005] [VSR-5.4-005] AOA endpoints MUST be mounted using
functions to enable Android userspace to communicate with AOA-compliant
accessories.

Change-Id: I67ae0a1051888b7c264571760faf312acfb1ab18